Move to using a string iterator and regex for parsing
I'm now using a std::string::iterator instead of a number to represent
the current location in str. This should be a bit better c++, I hope?.

I've also changed from my kind sloppy parsing to relying on regex as
well as adding support for parsing hex and decimal as well as dozenal
that was previously supported.